+++ src/term.c
@@ -361,6 +361,7 @@ term_reset(bool full)
     term.readline_mouse_1 = cfg.clicks_place_cursor;
     term.readline_mouse_2 = cfg.clicks_place_cursor;
     term.readline_mouse_3 = cfg.clicks_place_cursor;
+    term.emoji_width = false;
   }
 
   term.virtuallines = 0;